% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{advertising data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 200 rows and 5 columns.
}
\usage{
advertising
}
\description{
The dataframe consists of 200 observations with 5 columns on
advertising expenditures and sales (in 100s dollars) from 3
different media: TV, radio, and newspaper.

\tabular{ll}{
 \strong{V1} \tab numeric observation index\cr
 \strong{TV} \tab numeric expenditure for TV\cr
 \strong{radio} \tab numeric expenditure for radio\cr
 \strong{newspaper} \tab numeric expenditure for newspaper\cr
 \strong{sales} \tab numeric resulting sales
}
}
\keyword{datasets}
