% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloodpress}
\alias{bloodpress}
\title{bloodpress data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 8 columns.
}
\usage{
bloodpress
}
\description{
The data set of predictors originates
from [PennState Eberly College of Science](https://online.stat.psu.edu/stat501/lesson/welcome-stat-501)
Lesson 12: Multicollinearity & Other Regression Pitfalls.

The dataframe consists of 20 individuals with high blood pressure. Here
are the column measurements:

\tabular{ll}{
 \strong{Pt} \tab observation index \cr
 \strong{BP} \tab blood pressure , in mm Hg \cr
 \strong{Age} \tab age in years \cr
 \strong{Weight} \tab weight in kg \cr
 \strong{BSA} \tab body surface area in sq m \cr
 \strong{Dur} \tab duration of hypertension in years \cr
 \strong{Pulse} \tab basal pulse in beats per minute \cr
 \strong{Stress} \tab stress
}
}
\keyword{datasets}
