% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_correlatioins.R
\name{compute_correlations}
\alias{compute_correlations}
\title{Function computes a simple correlation between variables of a data frame}
\usage{
compute_correlations(df = NULL)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors values.}
}
\value{
A data frame of correlations between the variables.
}
\description{
The data frame should have observations along the rows without NA values
and the columns of variables are assumed to be numeric.
}
\author{
Rick Dean
}
