% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_report.R
\name{dataframe_report}
\alias{dataframe_report}
\title{Function provides a simple report on a dataframe destined to for
regression modeling, plotting, or other data investigations.}
\usage{
dataframe_report(df, miss_values = NULL)
}
\arguments{
\item{df}{The target dataframe from which some basic information on observed data
can be gathered.}

\item{miss_values}{A vector with numeric and character values that define in addition to \code{NA} and \code{NaN}, other values
considered as missing. Examples might be a vector with "na", "N/A", \code{999}.}
}
\value{
Returning a list with dataframes named:
\enumerate{
\item "missing",
\item "outliers"
\item "datatypes"
}
}
\description{
Function returns a named list of dataframes showing rows of outliers, missing
data, and types of variables from the target dataframe.
}
\author{
Rick Dean
}
