% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_beta.R
\name{mom_beta}
\alias{mom_beta}
\title{Function estimates the parameters of the beta distribution to replicate
a set of data based on their population mean and standard deviation.}
\usage{
mom_beta(n, mean, sd, min_val, max_val, digits = NULL, seed = NULL)
}
\arguments{
\item{n}{An integer that sets the number of beta distribution points to
estimate using stats::rbeta().}

\item{mean}{A numeric that sets the population mean of the observed data.}

\item{sd}{A numeric that sets the population standard deviation of the data.}

\item{min_val}{A numeric that sets the population's minimal value.}

\item{max_val}{A numeric that sets the population's maximum value.}

\item{digits}{An integer that sets the number of digits to round the returned
data.}

\item{seed}{An integer used by stats::rbeta() in setting the random seed.}
}
\value{
A list with the shape parameters used to estimate the beta distribution
along with data.frames of estimated data in both unscaled and scaled format.
}
\description{
The function estimates the shape parameters of the beta
distribution using the method of moments involving the submitted mean
and standard deviation.  The returned data set reflects the \link{0,1} range
of the beta distributions rescaled to the user's min/max units.
}
\author{
Rick Dean
}
