% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_BP_test_calc.R
\name{ols_BP_test_calc}
\alias{ols_BP_test_calc}
\title{Function computes the Breusch-Pagan Test for heteroskedasticity involving OLS models}
\usage{
ols_BP_test_calc(df = NULL, formula_obj = NULL)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}
}
\value{
Returning a data frame of F and LM and their p-values
}
\description{
Performs the BP-test from submitted lm formula and data frame. Function
implements both the F and LM versions of the BP test.
}
\author{
Rick Dean
}
