% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_F_calc.R
\name{ols_F_calc}
\alias{ols_F_calc}
\title{Function computes the linear F test based on the relative difference between
the sum of squared residuals in the general (unrestricted) model and the
restricted model.}
\usage{
ols_F_calc(
  df = NULL,
  ur_formula_obj = NULL,
  r_formula_obj = NULL,
  confid_level = 0.95
)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors}

\item{ur_formula_obj}{The unrestricted formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}

\item{r_formula_obj}{The restricted formula object following the rules of \code{stats::lm()} construction.}

\item{confid_level}{A numeric that sets the confidence level for computing the
critical value.}
}
\value{
Returns a named list with a data frame of computed F statistics along
with OLS coefficient estimates for each model.
}
\description{
Given formulas for the the unrestricted and restricted OLS models,
function returns a data frame with the computed F-statistic along with OLS coefficient
estimates for both models.
}
\author{
Rick Dean
}
