% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_matrix_calc.R
\name{ols_matrix_calc}
\alias{ols_matrix_calc}
\title{Function computes the OLS coefficients via matrix operations}
\usage{
ols_matrix_calc(X, Y)
}
\arguments{
\item{X}{A numeric matrix or data frame with rows of observations and their respective predictor
values across the columns.}

\item{Y}{A numeric matrix or data frame consisting of one column of the observations response
values.}
}
\value{
A named list with various OLS regression statistics
}
\description{
Function provides an alternative to \emph{stats::lm()} by using
matrix based arguments for predictors and response in estimating the OLS
parameters.
}
\author{
Rick Dean
}
