% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_regress_calc.R
\name{ols_regress_calc}
\alias{ols_regress_calc}
\title{Function computes linear OLS regression parameter estimates}
\usage{
ols_regress_calc(dep_str = NULL, data_df = NULL)
}
\arguments{
\item{dep_str}{A string that names the column in \code{data_df} containing the dependent
variable.}

\item{data_df}{A data frame containing values for both the observed independent
and dependent variables.}
}
\value{
Returning a named list of the linear OLS regression parameter estimates.
}
\description{
Function will calculate regression OLS parameters with input
from a data frame containing columns for the dependent and independent
variables. R based matrix techniques are used in the calculations.
}
\author{
Rick Dean
}
