% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilots}
\alias{pilots}
\title{pilots data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 40 rows and 7 columns.
}
\usage{
pilots
}
\description{
The dataframe consists of 40 observations with 7 columns.
Twenty engineer apprentices and twenty pilots were given six tests.
The tests measured the following attributes:

\enumerate{
  \item Intelligence
  \item Form Relations
  \item Dynamometer
  \item Dotting
  \item Sensory Motor Coordination
  \item Perservation
}

The data were obtained from the companion FTP site of the book Methods of
Multivariate Analysis by Alvin Renche.

The dataframe contains these 6 test scores plus a 7th column indicating
if the observation is an "Apprentice" or "Pilot".
}
\keyword{datasets}
