% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_distrib.R
\name{plot_beta_distrib}
\alias{plot_beta_distrib}
\title{Function estimates the parameters of the beta distribution to replicate
a set of data based on their population mean and standard deviation.}
\usage{
plot_beta_distrib(
  n = NULL,
  mean = NULL,
  sd = NULL,
  min_val = 0,
  max_val = 1,
  digits = NULL,
  seed = NULL,
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_title = NULL,
  bins = 100,
  bar_fill = NA,
  bar_color = "black",
  bar_alpha = 0.4,
  bar_size = 1
)
}
\arguments{
\item{n}{An integer that sets the number of beta distribution points to
estimate using stats::rbeta().}

\item{mean}{A numeric that sets the population mean of the observed data.}

\item{sd}{A numeric that sets the population standard deviation of the data.}

\item{min_val}{A numeric that sets the population's minimal value.}

\item{max_val}{A numeric that sets the population's maximum value.}

\item{digits}{An integer that sets the number of digits to round the returned
data.}

\item{seed}{An integer used by stats::rbeta() in setting the random seed.}

\item{title}{A string that sets the overall title.}

\item{subtitle}{A string that sets the overall subtitle.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{bins}{An integer that sets the number of bins for the histogram. Default is 100.}

\item{bar_fill}{A string that sets the fill color for the bars.}

\item{bar_color}{A string that sets the outline color for the bars.}

\item{bar_alpha}{A numeric that set the alpha component to \code{bar_color}.}

\item{bar_size}{A numeric that sets the outline thickness of the bars.}
}
\value{
A list with the shape parameters used to estimate the beta distribution
along with numeric vectors of estimated data in both unscaled and scaled format. Also
a ggplot2 histogram of the distribution is also provided.
}
\description{
The function estimates the shape parameters of the beta
distribution using the method of moments involving the submitted mean
and standard deviation.  The returned data set reflects the \link{0,1} range
of the beta distributions rescaled to the user's min/max units. A histogram
ggplot2 plot of the estimated distribution is also returned.
}
\author{
Rick Dean
}
