% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Function produces a heatmap plot of a correlation matrix.}
\usage{
plot_correlations(
  corr,
  cluster_corr = FALSE,
  tile_border_color = "white",
  label_sz = 11,
  label_color = "black",
  label_fontface = "plain",
  label_alpha = 1,
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  axis_text_size = 11,
  rot_x_tic_angle = 40,
  scale_limits = c(-1, 1),
  scale_colors = c("blue", "green"),
  mid_scale_color = "white",
  legend_title = ggplot2::waiver(),
  show_legend = TRUE
)
}
\arguments{
\item{corr}{A square, symmetric, named, numeric matrix of correlation values between -1 to 1 in value. This is typically
the Value from \code{stats::cor()}.}

\item{cluster_corr}{A logical that if TRUE will cluster the correlations across the heatmap scales.}

\item{tile_border_color}{A numeric that sets the heatmap tile's border color.}

\item{label_sz}{A numeric that sets the size of the label.}

\item{label_color}{A string that sets the label's color.}

\item{label_fontface}{A string that sets the label's font face. Acceptable values are \dQuote{plain}, \dQuote{bold},
\dQuote{italic}, \dQuote{bold.italic}.}

\item{label_alpha}{A numeric that sets the label's alpha value.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{rot_x_tic_angle}{A numeric that sets the angle of the x axis tic labels.}

\item{scale_limits}{A numeric vector of two elements giving the min/max values for the correlation scale. The default is
\code{c(-1.0, 1.0)}}

\item{scale_colors}{A character vector of two elements giving the low/high colors that are interpolated across \code{scale_limits}.}

\item{mid_scale_color}{A character string that sets the color for the mid value of \code{scale_limits}.}

\item{legend_title}{A character string that sets the legend's title.}

\item{show_legend}{A logical that controls the appearance of the legend.}
}
\value{
A ggplot2 plot object.
}
\description{
With the submission of a correlation matrix, the function returns a ggplot2 plot object that takes the form
of a color coded heatmap of the upper triangular portion of a symmetric correlation matrix.
}
\author{
Rick Dean
}
