% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_residuals.R
\name{plot_fit_residuals}
\alias{plot_fit_residuals}
\title{Function creates a scatter plot of the OLS fitted versus residual values.}
\usage{
plot_fit_residuals(
  df = NULL,
  formula_obj = NULL,
  id_col = NULL,
  label_threshold = NULL,
  label_color = "red",
  title = NULL,
  subtitle = NULL,
  x_title = "Fitted Values",
  y_title = "Residual Values",
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  x_log10 = FALSE,
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  y_log10 = FALSE,
  axis_text_size = 11,
  pts_color = "black",
  pts_fill = "white",
  pts_shape = 21,
  pts_stroke = 1,
  pts_alpha = 1,
  pts_size = 1,
  trend_line = TRUE,
  trend_line_color = "red",
  trend_line_size = 1,
  zero_line_color = "blue",
  zero_line_size = 1.4,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  do_x_title = TRUE,
  do_y_title = TRUE
)
}
\arguments{
\item{df}{A data frame containing values for both the predictor values and the associated
response variable.}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}

\item{id_col}{An optional argument that names the column from data frame \code{df} providing
each observation with a unique identification value.  If this argument is NULL then
data frame row numbers are used for identification.}

\item{label_threshold}{A numeric that sets the residual threshold beyond which observations
will be labeled with their id.}

\item{label_color}{A string that sets the label/point color for observations whose absolute
residual is greater than the \code{label_threshold}.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{x_title}{A string that sets the observed response x axis title.}

\item{y_title}{A string that sets the fitted response y axis title.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{x}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{Depending on the class of \code{x}, a numeric/Date/POSIXct vector or function that defines
the exact major tic locations along the x axis.}

\item{x_minor_breaks}{Depending on the class of \code{x}, a numeric/Date/POSIXct vector or function that defines
the exact minor tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as \code{x_major_breaks}, that labels the major tics.}

\item{x_log10}{A logical which if \code{TRUE} will use a log10 scale for the x axis.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use \code{NA} to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as \code{y_major_breaks}, that labels the major tics.}

\item{y_log10}{A logical which if \code{TRUE} will use a log10 scale for the y axis.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_shape}{A numeric integer that sets the shape of the points. Typical values are 21 \dQuote{circle},
22 \dQuote{square}, 23 \dQuote{diamond}, 24 \dQuote{up triangle}, 25 \dQuote{down triangle}.}

\item{pts_stroke}{A numeric that sets the drawing width for a point shape.}

\item{pts_alpha}{A numeric value that sets the alpha level of \code{pts_color}.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{trend_line}{A logical which if TRUE plots a polynomial based trend line across the residuals.}

\item{trend_line_color}{A string that sets the color of the trend line.}

\item{trend_line_size}{A numeric that sets the width of the trend line.}

\item{zero_line_color}{A string that sets the color of the zero horizontal reference line.}

\item{zero_line_size}{A numeric that sets the width of the zero horizontal reference line.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}
}
\value{
Function returns a ggplot2 object of fitted vs residual values.
}
\description{
Function returns a ggplot2 scatter plot object which can be further modified.
}
\author{
Rick Dean
}
