% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_scatter.R
\name{plot_matrix_scatter}
\alias{plot_matrix_scatter}
\title{Function produces a ggplot2 based matrix scatter plot.}
\usage{
plot_matrix_scatter(
  df,
  title = NULL,
  rot_y_tic_label = FALSE,
  plot_dim = 5.5,
  display_plot = TRUE
)
}
\arguments{
\item{df}{The target data frame from which the scatter plots are layed out in a
matrix like fashion.}

\item{title}{A string that sets the overall title.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for
enhanced readability.}

\item{plot_dim}{A numeric that sets the overall width/height of the plot in inches.}

\item{display_plot}{A logical that if TRUE will display the plot}
}
\value{
A grob object. Note that if \code{display_plot} is set to \code{FALSE}, then
to produce graphical output you must use \code{grid::grid.draw()} on the returned
grob object.
}
\description{
Function produces a scatter plot of all combinations of a submitted
data frame's variables.
}
\author{
Rick Dean
}
