% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_check.R
\name{plot_model_check}
\alias{plot_model_check}
\title{Function provides multi-plots to check model assumptions of heteroscedasticity, normally distributed
residuals and leverage of each datum.}
\usage{
plot_model_check(
  fitted_v,
  response_v,
  residual_v,
  id_v = NULL,
  residual_label_threshold = NULL,
  label_color = "red",
  label_sd = NULL,
  title = NULL,
  axis_text_size = 11,
  pts_color = "black",
  pts_fill = "white",
  pts_alpha = 1,
  pts_size = 1,
  histo_fill = "white",
  histo_alpha = 1,
  ref_line_color = "red",
  trend_line = TRUE,
  trend_line_color = "red"
)
}
\arguments{
\item{fitted_v}{A required numeric vector of fitted values.}

\item{response_v}{A required numeric vector of observed response/dependent values.}

\item{residual_v}{A required numeric vector of cooresponding residual values.}

\item{id_v}{An optional numeric/string vector that labels the fit/residual pairs. If this argument is NULL then
the fit/residual pairs are numbered for identification.}

\item{residual_label_threshold}{A numeric that sets the residual threshold beyond which observations
will be labeled with their id.}

\item{label_color}{A string that sets the label/point color for observations whose absolute
residual is greater than the \code{label_threshold}.}

\item{label_sd}{A numeric that sets the number times +/- residual standard deviations to plot as a
pair of horizontal dotted lines. Typical values could be 1 or 2 standard deviations.}

\item{title}{A string that sets the plot title.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_alpha}{A numeric value that sets the alpha level of \code{pts_color}.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{histo_fill}{A string that sets the histogram fill color.}

\item{histo_alpha}{A numeric that sets the histogram fill alpha.}

\item{ref_line_color}{A string that sets the reference lines color.}

\item{trend_line}{A logical which if TRUE plots a polynomial based trend line across the residuals.}

\item{trend_line_color}{A string that sets the color of the trend line.}
}
\description{
Function plots fitted values versus residuals, fitted values versus response
values, a QQ plot of the standardized residuals, and histogram of residual values.
}
\details{
For the QQ plot only standardized residuals are computed without taking the square root of the absolute value.
}
\author{
Rick Dean
}
