% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs_influence.R
\name{plot_obs_influence}
\alias{plot_obs_influence}
\title{Function creates a scatter plot of observation's row number versus its influence measure
in the OLS estimate.}
\usage{
plot_obs_influence(
  data_df = NULL,
  resp_str = NULL,
  influence_meas = "external",
  label_threshold = 3,
  label_color = "red",
  title = NULL,
  subtitle = NULL,
  x_title = "Observation ID",
  y_title = "Influence Value",
  rot_y_tic_label = FALSE,
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  axis_text_size = 11,
  pts_color = "black",
  pts_fill = "white",
  pts_shape = 21,
  pts_stroke = 1,
  pts_alpha = 1,
  pts_size = 1,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  do_x_title = TRUE,
  do_y_title = TRUE
)
}
\arguments{
\item{data_df}{A data frame containing values for both the observed predictors
and response variable.}

\item{resp_str}{A string that names the column in \code{data_df} containing the response
variable.}

\item{influence_meas}{A string that defines the type of influence measure to apply. Acceptable
values include \dQuote{internal}, \dQuote{external}, \dQuote{dffits}, and \dQuote{cook} for
internal/external studentized residuals, difference in fits, and Cook's distance respectively.}

\item{label_threshold}{A numeric that sets the measurement threshold beyond which observations
will be labelled with their id.}

\item{label_color}{A string that sets the label/point color for observations whose absolute
measurement is greater than the \code{label_threshold}.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{x_title}{A string that sets the observed response x axis title.}

\item{y_title}{A string that sets the fitted response y axis title.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use \code{NA} to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_shape}{A numeric integer that sets the shape of the points. Typical values are 21 \dQuote{circle},
22 \dQuote{square}, 23 \dQuote{diamond}, 24 \dQuote{up triangle}, 25 \dQuote{down triangle}.}

\item{pts_stroke}{A numeric that sets the drawing width for a point shape.}

\item{pts_alpha}{A numeric value that sets the alpha level of \code{pts_color}.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{do_x_title}{A logical that controls the appearance of the x axis title.}

\item{do_y_title}{A logical that controls the appearance of the y axis title.}
}
\value{
Function returns a ggplot2 object of observations versus influence measures
along with a data frame with columns for the observation id and influence measure.
}
\description{
Given the predictor and response data that defines an OLS model, an
estimate of each observation's influence on that model is plotted.  The measures of
influence include internal/external studentized residuals, difference in fits, and
Cook's distance.  The plot can be used to identify possible outlier data points.
}
\author{
Rick Dean
}
