% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Function produces a ggplot2 based scatter plot of all pairs of a dataframe's variables.}
\usage{
plot_pairs(
  df,
  var_name_scaling,
  title = NULL,
  title_sz = 14,
  axis_text_sz = 6,
  rot_y_tic_label = FALSE,
  plot_dim = 12
)
}
\arguments{
\item{df}{The required data frame from which the scatter plots' values are derived and laid out in a
matrix like fashion.}

\item{var_name_scaling}{A required named list where the name is the variable name from \code{df}
and the value is a vector defining the corresponding major axis scaling breaks. If the value
is NULL, then ggplot2 figures out the variable's scaling. The vector can be numeric/dates or
anything acceptable in scaling a ggplot2 axis.}

\item{title}{A string that sets the overall title.}

\item{title_sz}{A numeric that sets the title's font size. Default is 14.}

\item{axis_text_sz}{A numeric that sets the font size along the axis'. Default is 6.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for
enhanced readability.}

\item{plot_dim}{A numeric that sets the overall width/height of the plot in centimeters.}
}
\description{
In creating the \code{var_name_scaling} list, if we are looking at some regression model,
then suggest listing the response variable last.
}
\author{
Rick Dean
}
