% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_check.R
\name{plot_residuals_check}
\alias{plot_residuals_check}
\title{Function provides multi-plots to check model assumptions of heteroscedasticity, normally distributed
residuals and leverage of each datum.}
\usage{
plot_residuals_check(
  df = NULL,
  formula_obj = NULL,
  id_col = NULL,
  residual_label_threshold = NULL,
  leverage_label_threshold = NULL,
  label_color = "red",
  label_sd = NULL,
  title = NULL,
  axis_text_size = 11,
  pts_color = "black",
  pts_fill = "white",
  pts_alpha = 1,
  pts_size = 1,
  trend_line = TRUE,
  trend_line_color = "red",
  zero_line_color = "blue",
  zero_line_size = 1.4,
  display_plot = TRUE
)
}
\arguments{
\item{df}{A data frame containing values for both the predictor values and the associated
response variable.}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}

\item{id_col}{An optional argument that names the column from data frame \code{df} providing
each observation with a unique identification value.  If this argument is NULL then
data frame row numbers are used for identification.}

\item{residual_label_threshold}{A numeric that sets the residual threshold beyond which observations
will be labeled with their id.}

\item{leverage_label_threshold}{A numeric that sets the leverage threshold beyond which observations
will be labeled with their id.}

\item{label_color}{A string that sets the label/point color for observations whose absolute
residual is greater than the \code{label_threshold}.}

\item{label_sd}{A numeric that sets the number times +/- residual standard deviations to plot as a
pair of horizontal dotted lines. Typical values could be 1 or 2 standard deviations.}

\item{title}{A string that sets the plot title.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_alpha}{A numeric value that sets the alpha level of \code{pts_color}.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{trend_line}{A logical which if TRUE plots a polynomial based trend line across the residuals.}

\item{trend_line_color}{A string that sets the color of the trend line.}

\item{zero_line_color}{A string that sets the color of the zero horizontal reference line.}

\item{zero_line_size}{A numeric that sets the width of the zero horizontal reference line.}

\item{display_plot}{A logical that if TRUE will display the plot}
}
\value{
A grob object.
}
\description{
Function plots fitted values versus residuals, fitted values versus standardized
residuals, a QQ plot of the standardized residuals, and Cook's distance for a check on datum
leverage.
}
\details{
For the standardized residuals, the raw residuals are divided by their standard deviation.
The square root of the absolute value of each standardized residual is then plotted against their
respective fitted value.

For the QQ plot only standardized residuals are computed without taking the square root of the absolute value.
}
\author{
Rick Dean
}
