% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize}
\usage{
standardize(df = NULL, cols = NULL, mul = 1, do_na = TRUE)
}
\arguments{
\item{df}{A data frame with numeric columns to be standardized}

\item{cols}{A character vector of column names from \code{df} that are to be standardized.}

\item{mul}{An integer (typically 1 or 2) that multiplies the standard deviation
as the divisor for the z-score.}

\item{do_na}{A logical which if \code{TRUE} removes \code{NA} values in computing
the mean and standard deviation.}
}
\value{
Returns a data.table with the numeric columns standardized into z-scores.
}
\description{
Function computes the z-scores for numeric columns of a dataframe.

Scores are computed by dividing the raw differences from the
the mean by a multiple of the standard deviation.
}
