% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Therms18}
\alias{Therms18}
\title{Therms18 data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2500 rows and 2 columns.
}
\usage{
Therms18
}
\description{
The dataframe consists of 2500 observations with 2 columns
 on ratings for President Obama and President Trump. The ratings'
 scale range from 0 to 100 for unfavorable to favorable.

\tabular{ll}{
 \strong{ftobama} \tab measures for President Obama \cr
 \strong{fttrump} \tab measures for President Trump
}
}
\keyword{datasets}
