% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_predictor_collinearity.R
\name{low_predictor_collinearity}
\alias{low_predictor_collinearity}
\title{Function provides a way to identify a set of predictors with pairwise low collinearity among themselves.}
\usage{
low_predictor_collinearity(corr_matrix, threshold = 0.75)
}
\arguments{
\item{corr_matrix}{A symmetric numeric correlation matrix-like object with the predictor names for both the column names and row names.}

\item{threshold}{A numeric that sets the minimum correlation between pairs of predictors to run through the algorithm.}
}
\value{
Returning a named list with:
\enumerate{
\item \dQuote{predictors} A character vector with the names of predictors with pairwise low collinearity among themselves.
\item \dQuote{correlations} The correlation matrix with just the selected predictors.
\item \dQuote{max_correlation} The maximum correlation among all pairs of the selected predictors.
}
}
\description{
With the submission of a symmetric correlation matrix among predictor variables, the function will perform
the following algorithm:
\enumerate{
\item Create a starting list of the all the candidate predictors.
\item Create a second list of pairs of predictors with correlations above a given threshold and order the correlations
from high to low.
\item For each pair of predictors (call them A and B) in the ordered list, determine the average correlation between
predictor A and the other predictors.  Do the same for predictor B.
\item If A has a larger absolute average correlation, remove it from the ordered list and from the start list created in step 1;
otherwise remove predictor B.
\item Repeat steps 3-4 through the entire ordered list of correlations defined in step 2, removing potential predictors from the
starting list created in step 1.
\item The predictors left in the starting list are identified as having a low level of collinearity.
}
}
\author{
Rick Dean
}
