% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_BP_test_calc.R
\name{ols_BP_test_calc}
\alias{ols_BP_test_calc}
\title{Function computes the Breusch-Pagan Test for heteroskedasticity involving OLS models}
\usage{
ols_BP_test_calc(df = NULL, formula_obj = NULL, confid_level = 0.95)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}

\item{confid_level}{A numeric that defines the confidence level for estimating confidence
intervals. The default is 0.95.}
}
\value{
Returning a data frame with the standard F test of \code{lm} results and LM test and their p-values
}
\description{
Performs the BP-test from submitted lm formula and data frame. Function
implements both the F and LM versions of the BP test.
}
\details{
Background for creating this function was from "Using R for Introductory Econometrics" by
Florian Heiss, page 147.
}
\author{
Rick Dean
}
