% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_BP_test_calc.R
\name{ols_BP_test_calc}
\alias{ols_BP_test_calc}
\title{ols_BP_test_calc}
\usage{
ols_BP_test_calc(df = NULL, formula_obj = NULL)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
For example: \code{y ~ log(a) + b + I(b^2)}.}
}
\value{
Returning a list with the BP statistic and its p-value.
}
\description{
Function computes the Breusch-Pagan Test for heteroskedasticity involving OLS models

Performs the BP-test from submitted \code{lm} formula and data frame.

Background for creating this function was from "Using R for Introductory Econometrics" by
  Florian Heiss, page 147.
}
\examples{
library(wooldridge)
library(RregressPkg)

price_formula_obj = stats::as.formula("price~lotsize+sqrft+bdrms")
bp_lst <- RregressPkg::ols_BP_test_calc(
  df = wooldridge::hprice1,
  formula_obj = price_formula_obj
)

}
