% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_calc.R
\name{ols_calc}
\alias{ols_calc}
\title{Function computes linear OLS regression parameter estimates}
\usage{
ols_calc(df = NULL, formula_obj = NULL, na_omit = FALSE)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}

\item{na_omit}{A logical which if TRUE will omit rows that have NA values. If the
data contains NA's then the correlation matrix will not be computed.}
}
\value{
Returning a named list of data frames with OLS coefficient estimates, residual
statistics, ANOVA of the regression along with residuals, fitted values, and R^2.
}
\description{
The function requires a data frame with columns for observed response and predictors,
and a formula object (class = dQuote{formula}) that describes the OLS model of interest.
}
\author{
Rick Dean
}
