% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_calc.R
\name{ols_calc}
\alias{ols_calc}
\title{Function computes linear OLS regression parameter estimates}
\usage{
ols_calc(x = NULL, resp_col = NULL)
}
\arguments{
\item{x}{Either an object of class "lm" from an OLS estimate or a data frame containing
values for the predictor variables and the response variable.}

\item{resp_col}{A string that names the column from \code{x} containing the response
values if \code{x} is a data frame.}
}
\value{
Returning a named list of data frames with OLS coefficient estimates, residual
statistics, ANOVA of the regression along with residuals, fitted values, and R^2.
}
\description{
Function will calculate regression OLS parameters with input
from a data frame containing columns for a response variable and predictor
variables or a "lm" class object.
}
\author{
Rick Dean
}
