% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_calc.R
\name{ols_calc}
\alias{ols_calc}
\title{ols_calc}
\usage{
ols_calc(
  df = NULL,
  X_mt = NULL,
  Y_mt = NULL,
  formula_obj = NULL,
  confid_level = 0.95,
  na_omit = FALSE,
  print_detail = FALSE
)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors.}

\item{X_mt}{In lieu of both 'df' and 'formula_obj', a matrix of predictor values can be
submitted along with 'Y_mt'. Note that 'X_mt' must have column names. Also, to include
an intercept, 'X_mt' should have a column of 1's.}

\item{Y_mt}{In lieu of both 'df' and 'formula_obj', a single column matrix of response
values can be submitted along with 'X_mt'.}

\item{formula_obj}{A formula object following the rules of \code{stats::lm()} construction.
(e.g \code{y ~ log(a) + b + I(b^2))}. A constant parameter is implicitly added.}

\item{confid_level}{A numeric that defines the confidence level for estimating confidence
intervals. The default is 0.95.}

\item{na_omit}{A logical which if \code{TRUE} will omit rows that have \code{NA} values.}

\item{print_detail}{A logical which if \code{TRUE} will print a few statistics on the model.
The default is \code{FALSE}.}
}
\value{
Returning a named list of dataframes with OLS coefficient estimates, residual
 statistics, ANOVA of the regression along with residuals, fitted values, and R^2.
}
\description{
Function computes linear OLS regression parameter estimates

 The function requires a data frame with columns for observed response and predictors,
 and a formula object (class = "formula") that describes the OLS model of interest.
 Alternatively, the function will accept a matrix 'X_mt' and one column matrix 'Y_mt' for the
 predictor and response values respectively.
}
\examples{
library(wooldridge)
library(data.table)
library(RregressPkg)

gpa1_dt <- data.table::as.data.table(wooldridge::gpa1) |>
_[, skipped := -skipped] |>
  _[, .(colGPA, hsGPA, ACT, skipped)]

ols_ls <- RregressPkg::ols_calc(
  df = gpa1_dt,
  formula_obj = colGPA ~ hsGPA + ACT + skipped
)

}
