% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_F_calc.R
\name{ols_F_calc}
\alias{ols_F_calc}
\title{Function computes the linear F test based on the relative difference between
the sum of squared residuals in the general (unrestricted) model and the
restricted model.}
\usage{
ols_F_calc(
  df = NULL,
  ur_formula_obj = NULL,
  r_formula_obj = NULL,
  f_prob = 0.99
)
}
\arguments{
\item{df}{A data frame with columns for observed response and predictors}

\item{ur_formula_obj}{The unrestricted formula object following the rules of \code{stats::lm()} construction.
For example: y ~ log(a) + b + I(b^2) or suppress the constant with \code{0} in the formula.}

\item{r_formula_obj}{The restricted formula object following the rules of \code{stats::lm()} construction.}

\item{f_prob}{A numeric that sets the F probability for computing the
critical value.}
}
\value{
Returns a data frame with the computed F value and associated p value along
with sums of the squares for each model.
}
\description{

}
\author{
Rick Dean
}
