% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{plot_qq}
\usage{
plot_qq(
  df = NULL,
  numeric_col = NULL,
  id_col = NULL,
  standardize = FALSE,
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y_title = NULL,
  rot_y_tic_label = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_minor_breaks = waiver(),
  x_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  y_minor_breaks = waiver(),
  y_labels = waiver(),
  axis_text_size = 11,
  pts_color = "black",
  pts_fill = "white",
  pts_shape = 21,
  pts_stroke = 1,
  pts_alpha = 1,
  pts_size = 1,
  plot_ref_line = TRUE,
  ref_line_color = "red",
  ref_line_width = 1,
  plot_ci_lines = TRUE,
  ci_lines_color = "black",
  ci_lines_width = 1,
  ci_line_type = "solid",
  labels_n = NULL,
  label_color = "red"
)
}
\arguments{
\item{df}{A data frame with a column of numeric observations and optionally a column for
observation identifications.}

\item{numeric_col}{A string that names the numeric column from 'df' for QQ plotting.}

\item{id_col}{An optional argument that names the column from 'df' providing
each observation with a unique identification value.  If this argument is NULL then
row numbers of 'df' are used for identification.}

\item{standardize}{A logical which if \code{TRUE} will standardize the observations under \code{df$numeric_col}.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{x_title}{A string that sets the observed response x axis title.}

\item{y_title}{A string that sets the fitted response y axis title.}

\item{rot_y_tic_label}{A logical which if TRUE rotates the y tic labels 90 degrees for enhanced readability.}

\item{x_limits}{Depending on the class of \code{df$numeric_col}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis. Use \code{NA} to refer to the existing minimum and maximum.}

\item{x_major_breaks}{Depending on the class of \code{df$numeric_col}, a numeric/Date/POSIXct vector or function that defines
the exact major tic locations along the x axis.}

\item{x_minor_breaks}{Depending on the class of \code{df$numeric_col}, a numeric/Date/POSIXct vector or function that defines
the exact minor tic locations along the x axis.}

\item{x_labels}{A character vector with the same length as 'x_major_breaks', that labels the major tics.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis.
Use \code{NA} to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations along the y axis.}

\item{y_minor_breaks}{A numeric vector or function that defines the exact minor tic locations along the y axis.}

\item{y_labels}{A character vector with the same length as 'y_major_breaks', that labels the major tics.}

\item{axis_text_size}{A numeric that sets the font size along the axis'. Default is 11.}

\item{pts_color}{A string that sets the color of the points.}

\item{pts_fill}{A string that sets the fill color of the points.}

\item{pts_shape}{A numeric integer that sets the shape of the points. Typical values are 21 \dQuote{circle},
22 \dQuote{square}, 23 \dQuote{diamond}, 24 \dQuote{up triangle}, 25 \dQuote{down triangle}.}

\item{pts_stroke}{A numeric that sets the drawing width for a point shape.}

\item{pts_alpha}{A numeric value that sets the alpha level of \code{pts_color}.}

\item{pts_size}{A numeric value that sets the size of the points.}

\item{plot_ref_line}{A logical which if \code{TRUE} will plot a reference line.}

\item{ref_line_color}{A string that sets the color of the reference line.}

\item{ref_line_width}{A numeric that sets the reference line width.}

\item{plot_ci_lines}{A logical which if \code{TRUE} will plot upper/lower confidence lines.}

\item{ci_lines_color}{A string that sets the color of the confidence lines.}

\item{ci_lines_width}{A numeric that sets the confidence lines width.}

\item{ci_line_type}{A string that sets confidence lines line type. Acceptable values are \code{twodash, solid,
longdash, dotted, dotdash, dashed}}

\item{labels_n}{A numeric that sets the number of extreme values to label. If
\code{NULL} then no points are labeled.}

\item{label_color}{A string that sets the label color.}
}
\value{
Returning a ggplot2 object.
}
\description{
Function plots a quantile-quantile plot (QQ-plot) for comparing an
observed distribution with a normal distribution.
}
\examples{
library(ggplot2)
library(data.table)
library(RplotterPkg)
library(RregressPkg)

set.seed(20200825)
sample_obs_df <- data.frame(
  vals = stats::rnorm(20, 10, 3)
)

a_plot <- RregressPkg::plot_qq(
  df = sample_obs_df,
  numeric_col = "vals",
  title = "QQ plot of Random Observations",
  subtitle = "Sample is from a normal distribution",
  x_title = "Theoretical Normal Quantiles",
  y_title = "Sample Quantiles",
  rot_y_tic_label = TRUE,
  ci_lines_color = "blue",
  ci_line_type = "dashed",
  pts_size = 4,
  labels_n = 3
)

}
